requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.24.1
Pillow==9.4.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
